bug(modal)
fix bug with closing odalon succes
