+++ client/components/protected/MessageList.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { Modal } from 'react-bootstrap';
+import lodash from 'lodash';
 import MessageBox from './MessageBox';
-import AddMember from './AddMember';
 import Message from './Message';
 import AppStore from '../../stores/AppStore';
+import AppActions from '../../actions/AppActions';
 
 /**
 * A collection of message that displays the available messages
@@ -26,32 +26,109 @@ class MessageList extends React.Component {
     super(props);
     this.state = {
       isOpen: false,
+      userName: '',
+      userId: '',
+      error: '',
     };
-    this.closeModal = this.closeModal.bind(this);
-    this.openModal = this.openModal.bind(this);
+    this.openCreateGroupModal = this.openCreateGroupModal.bind(this);
+    this.onChange = this.onChange.bind(this);
+    this.onClick = this.onClick.bind(this);
+    this.changeToUserId = this.changeToUserId.bind(this);
   }
 
+
 /**
-* Handles close Modal event
+* @method onChange
+*
+* @description Monitors changes in the components and change the state
+*
+* @memberof MessageList
 *
 * @param {SyntheticEvent} event
 *
-* @returns {void} void
+* @returns {void}
 */
-  closeModal() {
-    this.setState({ isOpen: false });
+  onChange(event) {
+    this.setState({
+      userName: event.target.value
+    });
   }
 
+
 /**
-* Handles Open Modal event
-*
-* @param {SyntheticEvent} event
-*
-* @returns {void} void
+ * @description add a member to a group if a group is selected and
+ * username is not empty
+ *
+ * @param {SyntheticEvent} event
+ *
+ * @returns {void}
+ *
+ * @memberof MessageList
 */
-  openModal(event) {
+
+  onClick(event) {
     event.preventDefault();
-    this.setState({ isOpen: true });
+    if (!AppStore.getOpenedGroup()[0]) {
+      this.setState({
+        error: 'Kindly select a group first',
+      });
+      return true;
+    }
+    const user = {
+      userId: this.changeToUserId(this.state.userName),
+      userName: this.state.userName,
+      groupId: AppStore.getOpenedGroup()[0].groupId,
+    };
+    if (!user.userId) {
+      this.setState({
+        error: 'This User does not exist',
+      });
+    } else {
+      AppActions.addUserToGroup(user);
+      this.setState({
+        error: '',
+        userName: '',
+      });
+    }
+  }
+
+
+ /**
+   * @description describes a function that dynamically
+   * controls the behaviour of a modal
+   *
+   * @param { string } event
+   *
+   * @return { void }
+   *
+   * @memberof MessageList
+   */
+  openCreateGroupModal = (event) => {
+    event.preventDefault();
+    const $myModal = $('#modal');
+    $myModal.modal();
+  }
+
+
+/**
+ * @description function that get userid from username
+ *
+ * @param {String} userName
+ *
+ * @returns {String}
+ *
+ * @memberof MessageList
+ */
+  changeToUserId(userName) {
+    let userId;
+    lodash.map(AppStore.getUsers()).map((user) => {
+      if (userName === user.userName) {
+        userId = user.userId;
+      } else {
+        return null;
+      }
+    });
+    return userId;
   }
 
 /**
@@ -91,17 +168,57 @@ class MessageList extends React.Component {
             {groupName}
           </div>
           <div className="col-md-2">
-            <a onClick={this.openModal}> <b> Add new Member </b> </a>
+            <a onClick={this.openCreateGroupModal}> <b> Add new Member </b> </a>
           </div>
         </div>
-        <Modal show={this.state.isOpen} onHide={this.closeModal}>
-          <Modal.Body>
-            <AddMember />
-          </Modal.Body>
-          <Modal.Footer>
-            <a onClick={this.closeModal}> Close</a>
-          </Modal.Footer>
-        </Modal>
+        <div
+          className="modal fade modal2"
+          id="modal"
+          role="dialog"
+        >
+          <div className="modal-dialog">
+            <div className="modal-content">
+              <div className="modal-header">
+                <button
+                  type="button"
+                  className="close"
+                  data-dismiss="modal"
+                >
+                    &times;
+                      </button>
+                <h4 className="modal-title">
+                    Add Member
+                      </h4>
+                <div className="error"> {this.state.error} </div>
+              </div>
+              <div className="modal-body">
+                <input
+                  name="userName"
+                  onChange={this.onChange}
+                  value={this.state.userName}
+                  type="text"
+                  className="form-control"
+                  placeholder="Add member"
+                />
+              </div>
+              <div className="modal-footer">
+                <button
+                  type="button"
+                  className="btn btn-danger"
+                  data-dismiss="modal"
+                >
+                    Close
+                      </button>
+                <button
+                  onClick={this.onClick}
+                  type="submit"
+                  id="submit"
+                  className="btn btn-success"
+                > Submit </button>
+              </div>
+            </div>
+          </div>
+        </div>
         <div className="messages">
           {messageNodes}
         </div>
